chore: Module restructuring (so that reusable modules dependent on ftapi directory are together)
